refactor: 폴더 로직 분리
