+++ apps/web/app/review/[date]/page.tsx
@@ -1,9 +1,9 @@
-import { SpendingReview } from '@/widgets/spendingReview';
+import { Review } from '@/widgets/review';
 import React, { use } from 'react';
 
 type Props = { params: Promise<{ date: string }> };
 
 export default function ReviewPage({ params }: Props) {
   const { date } = use(params);
-  return <SpendingReview date={date} />;
+  return <Review date={date} />;
 }
