--- apps/web/src/features/spendingReview/index.ts
@@ -1,3 +1,0 @@
-export { useRetrospectExpenses } from './model/useRetrospectExpenses';
-export { patchRemind } from './api/patchRemind';
-export type { RemindItem } from './api/patchRemind';
