+++ apps/web/src/widgets/review/index.ts
@@ -1,0 +1,1 @@
+export { Review } from './ui/Review';
