+++ apps/web/src/widgets/review/ui/Review.css.ts
@@ -1,4 +1,4 @@
-import { radius, shadows, spacing, vars } from '@/shared/ui/theme.css';
+import { spacing } from '@/shared/ui/theme.css';
 import { keyframes, style } from '@vanilla-extract/css';
 
 const fadeSlideUp = keyframes({
@@ -29,36 +29,6 @@ export const container = style({
 
 export const titleText = style({
   marginLeft: '0.2rem',
-});
-
-export const reviewCardContainer = style({
-  display: 'flex',
-  flexDirection: 'column',
-  padding: `3.2rem ${spacing.xl}`,
-  gap: spacing.lg,
-  alignItems: 'flex-start',
-  backgroundColor: vars.color.bg.surface.secondary.default,
-  borderRadius: radius.lg,
-  boxShadow: shadows.shadow1,
-  width: '100%',
-  position: 'relative',
-});
-
-export const pageCount = style({
-  display: 'flex',
-  position: 'absolute',
-  top: spacing.xl,
-  right: spacing.xl,
-  padding: `${spacing.xs} ${spacing.sm2}`,
-  borderRadius: radius.full,
-  backgroundColor: vars.color.bg.neutral.subtle,
-  gap: '0.1rem',
-});
-
-export const textWrapper = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: spacing.xs,
 });
 
 export const pageIndicatorWrapper = style({
@@ -93,8 +63,4 @@ export const carouselSlide = style({
   padding: `0 ${spacing.xs2}`,
   marginTop: '1.25rem',
   boxSizing: 'border-box',
-});
-
-export const iconWrapper = style({
-  marginLeft: '-0.8rem',
 });
