+++ apps/web/src/widgets/review/ui/reviewCard/index.ts
@@ -1,0 +1,1 @@
+export { ReviewCard } from './ReviewCard';
