--- apps/web/src/widgets/spendingReview/index.ts
@@ -1,1 +1,0 @@
-export { SpendingReview } from './ui/SpendingReview';
